requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
lxml==4.9.1
nltk==3.7
Pillow==9.2.0
pycodestyle==2.8.0
Pygments==2.12.0
pypng==0.20220715.0
PyQRCode==1.2.1
pyzbar==0.1.9
qrcode==7.3.1
regex==2022.7.25
requests==2.28.1
rfc3986==1.5.0
rich==12.5.1
sniffio==1.2.0
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.11
